Ranking  * MacingCount Enum 제거 및 Ranking Enum으로 통합
