+++ src/main/java/lotto/domain/Ranking.java
@@ -1,5 +1,10 @@
 package lotto.domain;
 
+import java.util.Arrays;
+import java.util.Map;
+
+import static lotto.domain.RankPrintableFactory.makeRanksForPrint;
+
 public enum Ranking {
 
     FIRST_CLASS(1, 6, 2_000_000_000),
@@ -9,6 +14,8 @@ public enum Ranking {
     FIFTH_CLASS(5, 3, 5_000),
     ZERO_CLASS(0, 0, 0);
 
+    private static final int MINIMUM_MATCH_LOTTO = 3;
+    private static final int MATCHING_FIFTH_COUNT = 5;
     private int rank;
     private int countMatching;
     private int profit;
@@ -19,15 +26,36 @@ public enum Ranking {
         this.profit = profit;
     }
 
-    public int getRank() {
-        return this.rank;
+    public static Ranking valueOf(int countMatching) {
+        return Arrays.stream(Ranking.values())
+                .filter(e -> e.countMatching == countMatching)
+                .filter(f -> f.rank != 2)
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("Not Enum constant was found for matchCount : " + countMatching));
     }
 
-    public int getCountMatching() {
-        return this.countMatching;
+    public static int getProfit(Lottery lottery, WinnerLottery winnerLottery) {
+        return winnerLottery.rankLottery(lottery).profit;
     }
 
-    public int getProfit() {
-        return this.profit;
+    public static Ranking getRank(int count, boolean hasBonusNumber) {
+        if (count < MINIMUM_MATCH_LOTTO) {
+            return Ranking.ZERO_CLASS;
+        }
+
+        if(isMatchingFifth(count) && hasBonusNumber) {
+            return Ranking.SECOND_CLASS;
+        }
+
+        return Ranking.valueOf(count);
+    }
+
+    public static boolean isMatchingFifth(int count) {
+        return count == MATCHING_FIFTH_COUNT;
+    }
+
+    public void makeRanks(Map<String, Integer> ranks) {
+        RankPrintable rankPrintable = makeRanksForPrint(this.rank);
+        rankPrintable.makeRanksForPrint(ranks);
     }
 }
